[TASK] Fix phpstan checkFunctionArgumentTypes errors in ext:backend Form
This patch fixes incompatible type usage in function arguments
and is preparatory work for introducing native type hints and
strict mode in all core files.

Releases: master, 10.4
Resolves: #92579
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66158
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>

+++ typo3/sysext/backend/Classes/Form/FormDataProvider/DatabaseParentPageRow.php
@@ -36,7 +36,7 @@ class DatabaseParentPageRow extends AbstractDatabaseRecordProvider implements Fo
         if ($result['command'] === 'new') {
             if ($result['vanillaUid'] < 0) {
                 // vanillaUid points to a neighbor record in same table - get its record and its pid from there to find parent record
-                $neighborRow = $this->getRecordFromDatabase($result['tableName'], abs($result['vanillaUid']));
+                $neighborRow = $this->getRecordFromDatabase($result['tableName'], (int)abs($result['vanillaUid']));
                 if (!empty($neighborRow['t3ver_oid'])) {
                     $neighborRow = $this->getRecordFromDatabase($result['tableName'], (int)$neighborRow['t3ver_oid']);
                 }
